Move ClusterTestExtensionsUnitTest into test/junit
1. Move Move ClusterTestExtensionsUnitTest into test/junit into core/src/test/java/kafka/test/junit/ClusterTestExtensionsUnitTest.java
2. Modify processClusterTemplate modifier to protected
